# backend/CabBooking.Core/Models/Request/CabTypeRequestModel.cs
using System;
using System.Collections.Generic;
using System.Text;

namespace CabBooking.Core.Models.Request
{
    public class CabTypeRequestModel
    {
        public int Id { get; set; }
        public String Name { get; set; }
    }
}