# backend/CabBooking.Core/Models/Request/PlaceRequestModel.cs
using System;
using System.Collections.Generic;
using System.Text;

namespace CabBooking.Core.Models.Request
{
    public class PlaceRequestModel
    {
        public int Id { get; set; }
        public String Name { get; set; }
    }
}