# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CabBooking.API",
    "version": "v1"
  },
  "paths": {
    "/api/Booking/CabBooking/{id}": {
      "get": {
        "tags": [
          "Booking"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Booking/list": {
      "get": {
        "tags": [
          "Booking"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Booking/add": {
      "post": {
        "tags": [
          "Booking"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Booking/delete/{id}": {
      "delete": {
        "tags": [
          "Booking"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Booking/update": {
      "put": {
        "tags": [
          "Booking"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BookingHistory/list": {
      "get": {
        "tags": [
          "BookingHistory"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BookingHistory/add": {
      "post": {
        "tags": [
          "BookingHistory"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingHistoryRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingHistoryRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BookingHistoryRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BookingHistory/delete/{id}": {
      "delete": {
        "tags": [
          "BookingHistory"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BookingHistory/update": {
      "put": {
        "tags": [
          "BookingHistory"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingHistoryRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingHistoryRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BookingHistoryRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CabType/list": {
      "get": {
        "tags": [
          "CabType"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CabType/add": {
      "post": {
        "tags": [
          "CabType"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CabTypeRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CabTypeRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CabTypeRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CabType/delete/{id}": {
      "delete": {
        "tags": [
          "CabType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/CabType/update": {
      "put": {
        "tags": [
          "CabType"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CabTypeRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CabTypeRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CabTypeRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Place/list": {
      "get": {
        "tags": [
          "Place"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Place/add": {
      "post": {
        "tags": [
          "Place"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Place/delete/{id}": {
      "delete": {
        "tags": [
          "Place"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Place/update": {
      "put": {
        "tags": [
          "Place"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BookingRequestModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "fromPlaceId": {
            "type": "integer",
            "format": "int32"
          },
          "toPlaceId": {
            "type": "integer",
            "format": "int32"
          },
          "pickupAddress": {
            "type": "string",
            "nullable": true
          },
          "cabTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "contactNo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BookingHistoryRequestModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "fromPlaceId": {
            "type": "integer",
            "format": "int32"
          },
          "toPlaceId": {
            "type": "integer",
            "format": "int32"
          },
          "pickupAddress": {
            "type": "string",
            "nullable": true
          },
          "cabTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "contactNo": {
            "type": "string",
            "nullable": true
          },
          "charge": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "feedback": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CabTypeRequestModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlaceRequestModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}